# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Foros, Mensajes y Respuestas",
        "description": "Documentación de la API para la gestión de foros, mensajes, respuestas y autenticación.",
        "contact": {
            "email": "1813370@alu.murciaeduca.es"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://catalogovideojuegos.test/api",
            "description": "Servidor Principal de la API"
        }
    ],
    "paths": {
...
        "/foros/{foro}/mensajes": {
            "get": {
                "tags": [
                    "Mensajes de Foro"
                ],
                "summary": "Obtener mensajes de un foro específico",
                "operationId": "getMensajesForo",
                "parameters": [
                    {
                        "name": "foro",
                        "in": "path",
                        "description": "ID del foro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Paginado",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MensajeForo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Foro no encontrado"
                    }
                }
            },
            "post": {
                "tags": [
                    "Mensajes de Foro"
                ],
                "summary": "Crear un nuevo mensaje en un foro",
                "operationId": "storeMensajeForo",
                "parameters": [
                    {
                        "name": "foro",
                        "in": "path",
                        "description": "ID del foro donde crear el mensaje",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Contenido del mensaje",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MensajeForoStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mensaje creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MensajeForo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado para crear mensajes en este foro (si aplica policy)"
                    },
                    "404": {
                        "description": "Foro no encontrado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/mensajes/{mensajeForo}": {
            "get": {
                "tags": [
                    "Mensajes de Foro"
                ],
                "summary": "Obtener un mensaje específico de un foro",
                "description": "Retorna los datos de un mensaje de foro.",
                "operationId": "getMensajeForoById",
                "parameters": [
                    {
                        "name": "mensajeForo",
                        "in": "path",
                        "description": "ID del mensaje de foro",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MensajeForo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mensaje no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mensajes de Foro"
                ],
                "summary": "Actualizar un mensaje de foro existente",
                "operationId": "updateMensajeForo",
                "parameters": [
                    {
                        "name": "mensajeForo",
                        "in": "path",
                        "description": "ID del mensaje a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del mensaje a actualizar. 'contenido' es el campo principal.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MensajeForoStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mensaje actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MensajeForo"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Mensaje no encontrado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Mensajes de Foro"
                ],
                "summary": "Eliminar un mensaje de foro",
                "operationId": "deleteMensajeForo",
                "parameters": [
                    {
                        "name": "mensajeForo",
                        "in": "path",
                        "description": "ID del mensaje a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensaje eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Mensaje eliminado correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Mensaje no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserLite": {
                "title": "User Lite Information",
                "description": "Información básica del usuario",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Nombre Usuario"
                    }
                },
                "type": "object"
            },
...
            "RespuestaForo": {
                "title": "Respuesta de Mensaje de Foro",
                "description": "Representación de una respuesta a un mensaje de foro",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 101
                    },
                    "contenido": {
                        "type": "string",
                        "example": "Esta es una respuesta."
                    },
                    "usuario": {
                        "$ref": "#/components/schemas/UserLite"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T10:30:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T10:35:00.000000Z"
                    }
                },
                "type": "object"
            },
            "MensajeForo": {
                "title": "Mensaje de Foro",
                "description": "Representación de un mensaje dentro de un foro",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "contenido": {
                        "type": "string",
                        "example": "Contenido del mensaje principal."
                    },
                    "usuario": {
                        "$ref": "#/components/schemas/UserLite"
                    },
                    "respuestas": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RespuestaForo"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T10:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-04T10:05:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "ErrorValidation": {
                "title": "Error de Validación",
                "description": "Objeto de error devuelto en caso de fallos de validación (422)",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "Objeto con los errores de validación por campo",
                        "type": "object",
                        "example": {
                            "titulo": [
                                "El campo título es obligatorio."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "MensajeForoStoreRequest": {
                "title": "Mensaje Foro Store Request Body",
                "description": "Cuerpo de la petición para crear un mensaje de foro",
                "required": [
                    "contenido"
                ],
                "properties": {
                    "contenido": {
                        "type": "string",
                        "example": "Este es mi nuevo mensaje."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación por Token Bearer (Sanctum)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}